View functionality.
Implemented a working ViewTask page that allows the user to view the details of the task selected with options to edit and delete the task. Updated the App.jsx, TaskActions.jsx, TaskItem.jsx and Dashboard.jsx to handle this new logic. Added new view buttons to the dashboard and task list page.