UI and responsiveness changes.
Moved the README file to be at the root of the project to display the setup instructions correctly. Implemented a burger menu at smaller screens for the buttons in the layout component. Updated the dashboard and tasks views to have the action buttons in a drop down menu on smaller screens to save screen space and to help the layout of the task cards.